Fix a couple of warnings on Windows
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@881440 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/internal/util/HexStringParser.h
@@ -47,13 +47,13 @@ namespace util{
 
         int EXPONENT_WIDTH;
         int MANTISSA_WIDTH;
-        long EXPONENT_BASE;
-        long MAX_EXPONENT;
-        long MIN_EXPONENT;
-        long MANTISSA_MASK;
-        long sign;
-        long exponent;
-        long mantissa;
+        long long EXPONENT_BASE;
+        long long MAX_EXPONENT;
+        long long MIN_EXPONENT;
+        long long MANTISSA_MASK;
+        long long sign;
+        long long exponent;
+        long long mantissa;
         std::string abandonedNumber;
 
     public:
